backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
cohere>=5.0.0
qdrant-client>=1.9.0
python-pptx>=0.6.23
python-docx>=1.1.0
pypdf>=4.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
openai>=1.0.0